modified auth

+++ src/components/QuotesList.jsx
@@ -7,6 +7,9 @@ import Quote from './ListItem/Quote'
 
 import { deleteQuote } from '../redux/quote/quote.actions'
 
+import {useAuthStatus} from '../Hooks/useAuthStatus'
+import { useNavigate, useLocation } from 'react-router-dom';
+
 import Modal from 'react-modal';
 const appElement = document.getElementById('content');
 Modal.setAppElement(appElement);
@@ -24,8 +27,6 @@ export const customStyles = {
         },
     };
 
-
-
 const QuotesList = () => {
     // const [modalIsOpen, setIsOpen] = useState(false);
     
@@ -42,18 +43,40 @@ const QuotesList = () => {
 
     // const quoteLists = useSelector(state => state.quoteLists)
 
+    const {loggedIn} = useAuthStatus()
+    const location = useLocation()
+    const navigate = useNavigate()
+
     const dispatch = useDispatch();
     
     const deleteHandler = (id) => {
-        if(window.confirm('Are you sure?')){
-            dispatch(deleteQuote(id))
+        if(loggedIn) {
+            if(window.confirm('Are you sure?')){
+                dispatch(deleteQuote(id))
+            }
+        } else {
+            console.log(location)
+            navigate('/Login', { from : location})
         }
     }
+
+    const favHandler = (id) => {
+        if(loggedIn) {
+            console.log('fave!')
+            // setIsOpen(true);
+        } else {
+            console.log(location)
+            navigate('/Login', { from : location})
+        }
+    } 
+
+
 
     return (
         <div className="container mx-auto">
             <Quote 
                 // openModal={openModal}
+                favHandler={favHandler}
                 deleteHandler={deleteHandler}
             >
             </Quote>
